Move time_block to utils
